/* my oh my, it seems that there is a lot to do and this is quite a big comment.
Added the container-case to the commands look, get, drop, search, give, and read.

Also fixed some small faults where I ordered by itemid instead of containerid, in both
inventory as well as roomdescription.

One of the things that is quite convenient is that a containerid always means
that the amount=1. Which means that in most cases a simple update statement
is enough to, for example, drop a container or 'give' a container.

In the case where containerid =0, it can just be handled as a common item.

Also removed some nasty debugging statements for sql.
*/