/* Changed commandlineinterface to the standard #ifdef DEBUG in enter.c, mudrelogin.c, and newchar.c. Removed things like "DebugOptionsOn" and commandlineinterface = 0.
These Pre-Processor directives have the advantage that they can be
provided during a compile statement. For instance:

cc file.c -DDEBUG

Would define the DEBUG directive, effectively turning on debugging
in the C files. I still have to think of a way to get it to work with my
Makefile though.

It has the added advantage that debug code will not be included in the
compilation if not specifically called for, which means the code
will run faster without debug.
*/