/* replaced all the nice tokens[x] and aantal calls with four functions in typedefs.c namely: setTokenAmount(int) getTokenAmount() setTokens(char **) getToken(int)
This makes it easier for me to check wether or not the requested
tokens goes beyond the range of available tokens. Besides,
it is much more neato

Also performed some checks in some of the cases to make sure
that the amount of tokens available is of the correct amount.

Also removed the fight command until more info available on fighting,
and added the drop command. Erased it somehow somewhere.  */