/* fixed a little '#ifdef DEBUG' item, no biggy.
What IS a biggy is the thing I fixed that also froze the game.
Apparently entering "spaces" (any number) would freeze the game.
Seeing, as the 'command' variable is copied to the 'junk' variable
and all extraneous spaces are removed from the 'junk' variable, it
stands to reason that it is much better to compare 'junk' with '\0'
then command in order to fix the spaces thing.

Just a note: never ever change 'command'. You never know if the game
might provide a 'constant' to the procedure which would result in a
SegFault.
*/