/* removed and added and changed the "include" directives in both the header files as well as the source files.
The reason for this was the fact that I finally figured out how it 'should'
work instead of the recursive header file format that I've been using.

The principle is quite simple:
- do not put #include directives into header files (.h), because in header files
no source code is required and therefore libraries are not neeeded.
The only reason to add it is, for instance in the case of mysql.h, to make
use of typedefs of new variables.
- put all necessary #include directives in the source files (.c), and ONLY the
ones of which you are certain that the commands in that header file
are actually used in the source code.

To be more precise: the header files are strictly used in order for the
compiler to be able to 'compile' the source code (.c) files and by
using the 'headers' of methods that are defined someplace else.
*/